feat: start JOI

+++ controllers/cards.js
@@ -1,14 +1,17 @@
 const Card = require('../models/card');
 const ErrorNotFound = require('../errors/ErrorNotFound');
+const ErrorForbidden = require('../errors/ErrorForbidden');
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
   Card.create({ name, link, owner })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
-      if (err.name === 'ValidationError') {
+      next(err);
+
+      /*if (err.name === 'ValidationError') {
         let errorMessage = 'В следующих полях введены неверные данные: ';
         const errorValues = Object.values(err.errors);
         errorValues.forEach((errVal) => {
@@ -17,36 +20,40 @@ module.exports.createCard = (req, res) => {
           }
         });
         res.status(400).send({ message: errorMessage });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
-      }
+      }*/
     });
 };
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch((err) => {
+      next(err);
+    });
 };
 
-module.exports.delCardById = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.delCardById = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .orFail(() => {
       throw new ErrorNotFound('Карточки с таким ID не существует');
     })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.statusCode === 404) {
-        res.status(404).send({ message: err.errorMessage });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Неверно введён ID' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
+    .then((card) => {
+      if (card.owner.toString() !== req.user._id) {
+        throw new ErrorForbidden('Вы не можете удалять чужие карточки');
       }
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((card) => res.send({ data: card }))
+        .catch((err) => {
+          next(err);
+
+          /*if (err.name === 'CastError') {
+            res.status(400).send({ message: 'Неверно введён ID' });
+          }*/
+        });
     });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -57,17 +64,15 @@ module.exports.likeCard = (req, res) => {
     })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
-      if (err.statusCode === 404) {
-        res.status(404).send({ message: err.errorMessage });
-      } else if (err.name === 'CastError') {
+      next(err);
+
+      /*if (err.name === 'CastError') {
         res.status(400).send({ message: 'Неверно введён ID' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
-      }
+      }*/
     });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -78,12 +83,10 @@ module.exports.dislikeCard = (req, res) => {
     })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
-      if (err.statusCode === 404) {
-        res.status(404).send({ message: err.errorMessage });
-      } else if (err.name === 'CastError') {
+      next(err);
+
+      /*if (err.name === 'CastError') {
         res.status(400).send({ message: 'Неверно введён ID' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
-      }
+      }*/
     });
 };
